# 1-analytics/api/openapi.yaml
openapi: 3.0.3
info:
  title: Vertical Climbing Client API
  description: |
    OpenAPI specification for the client mobile application of the Vertical climbing gym.
    Covers registration, viewing training slots, booking, cancellation, notifications,
    visiting rules, and optional instructor rating.
  version: 0.1.0
servers:
  - url: https://api.vertical-climbing.local/api/v1
    description: Local/staging environment
security:
  - bearerAuth: []
tags:
  - name: Auth
    description: Client authentication and registration
  - name: Slots
    description: Training slots and schedule browsing
  - name: Bookings
    description: Booking creation, listing and cancellation
  - name: Notifications
    description: Notification settings for reminders and cancellations
  - name: Info
    description: Visiting rules and general information
  - name: Ratings
    description: Instructor feedback (post-MVP)
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new client
      operationId: registerClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Client registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input or age validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in an existing client
      operationId: loginClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /info/visiting-rules:
    get:
      tags: [Info]
      summary: Get the visiting rules document
      operationId: getVisitingRules
      responses:
        '200':
          description: Visiting rules returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitingRules'
  /slots:
    get:
      tags: [Slots]
      summary: List available training slots
      operationId: listSlots
      parameters:
        - in: query
          name: fromDate
          schema:
            type: string
            format: date
          description: Start of date range
        - in: query
          name: toDate
          schema:
            type: string
            format: date
          description: End of date range
        - in: query
          name: formatId
          schema:
            type: string
        - in: query
          name: instructorId
          schema:
            type: string
        - in: query
          name: timeOfDay
          schema:
            type: string
            enum: [morning, day, evening]
        - in: query
          name: onlyAvailable
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Slot list returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TrainingSlot'
                  total:
                    type: integer
                    example: 12
  /slots/{slotId}:
    get:
      tags: [Slots]
      summary: Get details of a specific slot
      operationId: getSlotById
      parameters:
        - in: path
          name: slotId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Slot details returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingSlot'
        '404':
          description: Slot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings:
    get:
      tags: [Bookings]
      summary: List bookings of the current client
      operationId: listBookings
      responses:
        '200':
          description: Bookings returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'
    post:
      tags: [Bookings]
      summary: Create a booking for a training slot
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '409':
          description: No available seats due to conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Slot is cancelled or no longer available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings/{bookingId}:
    get:
      tags: [Bookings]
      summary: Get a booking by id
      operationId: getBookingById
      parameters:
        - in: path
          name: bookingId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Booking details returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Bookings]
      summary: Cancel a booking
      operationId: cancelBooking
      parameters:
        - in: path
          name: bookingId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Booking cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '403':
          description: Cancellation not allowed due to less than 24 hours before start
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me/notification-settings:
    get:
      tags: [Notifications]
      summary: Get notification settings for the current client
      operationId: getNotificationSettings
      responses:
        '200':
          description: Notification settings returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationSettings'
    put:
      tags: [Notifications]
      summary: Update notification settings
      operationId: updateNotificationSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationSettings'
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationSettings'
  /ratings:
    post:
      tags: [Ratings]
      summary: Submit an instructor rating (post-MVP)
      operationId: createRating
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRatingRequest'
      responses:
        '201':
          description: Rating created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstructorRating'
        '400':
          description: Invalid input or rating window expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
    RegisterRequest:
      type: object
      required: [email, password, birthDate, acceptSafetyRules]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        birthDate:
          type: string
          format: date
        acceptSafetyRules:
          type: boolean
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: bearer
        client:
          $ref: '#/components/schemas/Client'
    Client:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        registeredAt:
          type: string
          format: date-time
    TrainingSlot:
      type: object
      properties:
        id:
          type: string
        startTime:
          type: string
          format: date-time
        durationMinutes:
          type: integer
          example: 90
        capacity:
          type: integer
          example: 16
        freeSeats:
          type: integer
          example: 8
        status:
          type: string
          enum: [scheduled, cancelled, full]
        instructor:
          $ref: '#/components/schemas/Instructor'
        format:
          $ref: '#/components/schemas/TrainingFormat'
        zone:
          $ref: '#/components/schemas/Zone'
        rentalPrice:
          type: number
          format: float
          example: 7.5
      required: [id, startTime, durationMinutes, capacity, freeSeats, status]
    Instructor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        specialization:
          type: string
    TrainingFormat:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        level:
          type: string
          enum: [beginner, advanced]
    Zone:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        maxCapacity:
          type: integer
    CreateBookingRequest:
      type: object
      required: [slotId, seatsCount, rentalOption]
      properties:
        slotId:
          type: string
        seatsCount:
          type: integer
          minimum: 1
        rentalOption:
          type: string
          enum: [own, rental]
    Booking:
      type: object
      properties:
        id:
          type: string
        slot:
          $ref: '#/components/schemas/TrainingSlot'
        seatsCount:
          type: integer
        status:
          type: string
          enum: [active, cancelled_by_client, cancelled_by_gym]
        createdAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
        cancellationReason:
          type: string
        equipmentRental:
          $ref: '#/components/schemas/EquipmentRental'
      required: [id, slot, seatsCount, status, createdAt]
    EquipmentRental:
      type: object
      properties:
        id:
          type: string
        useRental:
          type: boolean
        rentalPrice:
          type: number
          format: float
    NotificationSettings:
      type: object
      properties:
        enabled:
          type: boolean
        reminderMinutesBefore:
          type: integer
          example: 60
        cancellationNotifications:
          type: boolean
    VisitingRules:
      type: object
      properties:
        title:
          type: string
        address:
          type: string
        schedule:
          type: string
        whatToBring:
          type: array
          items:
            type: string
        safetyRules:
          type: array
          items:
            type: string
    CreateRatingRequest:
      type: object
      required: [bookingId, score, reviewText]
      properties:
        bookingId:
          type: string
        score:
          type: integer
          minimum: 1
          maximum: 5
        reviewText:
          type: string
    InstructorRating:
      type: object
      properties:
        id:
          type: string
        bookingId:
          type: string
        score:
          type: integer
        reviewText:
          type: string
        createdAt:
          type: string
          format: date-time